extended grammar and fixed gperf error

+++ include/tokenizer.h
@@ -67,6 +67,6 @@ bool is_number(char c);
 bool is_alpha(char c);
 void print_token(Token tok);
 TokenArray tokenize(const char* buf, size_t buf_size);
-KeywordEntry *lookup_keyword(const char *str, unsigned int len);
+KeywordEntry *lookup_keyword(const char *str, size_t len);
 
 #endif
